added trace-file backup
