+++ src/Test/Perf/Runner/Tools.cs
@@ -59,55 +59,6 @@ namespace Roslyn.Test.Performance.Runner
             }
         }
 
-
-        public static void UploadTraces(string sourceFolderPath, string destinationFolderPath)
-        {
-            RuntimeSettings.logger.Log("Uploading traces");
-            if (Directory.Exists(sourceFolderPath))
-            {
-                var directoriesToUpload = new DirectoryInfo(sourceFolderPath).GetDirectories("DataBackup*");
-                if (directoriesToUpload.Count() == 0)
-                {
-                    RuntimeSettings.logger.Log($"There are no trace directory starting with DataBackup in {sourceFolderPath}");
-                    return;
-                }
-
-                var perfResultDestinationFolderName = string.Format("PerfResults-{0:yyyy-MM-dd_hh-mm-ss-tt}", DateTime.Now);
-
-                var destination = Path.Combine(destinationFolderPath, perfResultDestinationFolderName);
-                foreach (var directoryToUpload in directoriesToUpload)
-                {
-                    var destinationDataBackupDirectory = Path.Combine(destination, directoryToUpload.Name);
-                    if (Directory.Exists(destinationDataBackupDirectory))
-                    {
-                        Directory.CreateDirectory(destinationDataBackupDirectory);
-                    }
-
-                    CopyDirectory(directoryToUpload.FullName, RuntimeSettings.logger, destinationDataBackupDirectory);
-                }
-
-                foreach (var file in new DirectoryInfo(sourceFolderPath).GetFiles().Where(f => f.Name.StartsWith("ConsumptionTemp", StringComparison.OrdinalIgnoreCase) || f.Name.StartsWith("Roslyn-", StringComparison.OrdinalIgnoreCase)))
-                {
-                    File.Copy(file.FullName, Path.Combine(destination, file.Name));
-                }
-            }
-            else
-            {
-                RuntimeSettings.logger.Log($"sourceFolderPath: {sourceFolderPath} does not exist");
-            }
-        }
-
-        public static void CopyDirectory(string source, ILogger logger, string destination, string argument = @"/mir")
-        {
-            var result = ShellOut("Robocopy", $"{argument} {source} {destination}", workingDirectory: "");
-
-            // Robocopy has a success exit code from 0 - 7
-            if (result.Code > 7)
-            {
-                throw new IOException($"Failed to copy \"{source}\" to \"{destination}\".");
-            }
-        }
-
         public static string FirstLine(string input)
         {
             return input.Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None)[0];
